This change moves plan attribute resolution to its own component
